refactor(server): system config
