+++ server/src/entities/system-config.entity.ts
@@ -1,4 +1,4 @@
-import { ConcurrentQueueName } from 'src/interfaces/job.interface';
+import { SystemConfig } from 'src/config';
 import { Column, Entity, PrimaryColumn } from 'typeorm';
 
 export type SystemConfigValue = string | string[] | number | boolean;
@@ -142,198 +142,4 @@ export class SystemConfigEntity<T = SystemConfigValue> {
 
   @Column({ type: 'varchar', nullable: true, transformer: { to: JSON.stringify, from: JSON.parse } })
   value!: T;
-}
-
-export enum TranscodePolicy {
-  ALL = 'all',
-  OPTIMAL = 'optimal',
-  BITRATE = 'bitrate',
-  REQUIRED = 'required',
-  DISABLED = 'disabled',
-}
-
-export enum TranscodeTarget {
-  NONE,
-  AUDIO,
-  VIDEO,
-  ALL,
-}
-
-export enum VideoCodec {
-  H264 = 'h264',
-  HEVC = 'hevc',
-  VP9 = 'vp9',
-  AV1 = 'av1',
-}
-
-export enum AudioCodec {
-  MP3 = 'mp3',
-  AAC = 'aac',
-  LIBOPUS = 'libopus',
-}
-
-export enum TranscodeHWAccel {
-  NVENC = 'nvenc',
-  QSV = 'qsv',
-  VAAPI = 'vaapi',
-  RKMPP = 'rkmpp',
-  DISABLED = 'disabled',
-}
-
-export enum ToneMapping {
-  HABLE = 'hable',
-  MOBIUS = 'mobius',
-  REINHARD = 'reinhard',
-  DISABLED = 'disabled',
-}
-
-export enum CQMode {
-  AUTO = 'auto',
-  CQP = 'cqp',
-  ICQ = 'icq',
-}
-
-export enum Colorspace {
-  SRGB = 'srgb',
-  P3 = 'p3',
-}
-
-export enum ImageFormat {
-  JPEG = 'jpeg',
-  WEBP = 'webp',
-}
-
-export enum LogLevel {
-  VERBOSE = 'verbose',
-  DEBUG = 'debug',
-  LOG = 'log',
-  WARN = 'warn',
-  ERROR = 'error',
-  FATAL = 'fatal',
-}
-
-export interface SystemConfig {
-  ffmpeg: {
-    crf: number;
-    threads: number;
-    preset: string;
-    targetVideoCodec: VideoCodec;
-    acceptedVideoCodecs: VideoCodec[];
-    targetAudioCodec: AudioCodec;
-    acceptedAudioCodecs: AudioCodec[];
-    targetResolution: string;
-    maxBitrate: string;
-    bframes: number;
-    refs: number;
-    gopSize: number;
-    npl: number;
-    temporalAQ: boolean;
-    cqMode: CQMode;
-    twoPass: boolean;
-    preferredHwDevice: string;
-    transcode: TranscodePolicy;
-    accel: TranscodeHWAccel;
-    tonemap: ToneMapping;
-  };
-  job: Record<ConcurrentQueueName, { concurrency: number }>;
-  logging: {
-    enabled: boolean;
-    level: LogLevel;
-  };
-  machineLearning: {
-    enabled: boolean;
-    url: string;
-    clip: {
-      enabled: boolean;
-      modelName: string;
-    };
-    facialRecognition: {
-      enabled: boolean;
-      modelName: string;
-      minScore: number;
-      minFaces: number;
-      maxDistance: number;
-    };
-  };
-  map: {
-    enabled: boolean;
-    lightStyle: string;
-    darkStyle: string;
-  };
-  reverseGeocoding: {
-    enabled: boolean;
-  };
-  oauth: {
-    autoLaunch: boolean;
-    autoRegister: boolean;
-    buttonText: string;
-    clientId: string;
-    clientSecret: string;
-    defaultStorageQuota: number;
-    enabled: boolean;
-    issuerUrl: string;
-    mobileOverrideEnabled: boolean;
-    mobileRedirectUri: string;
-    scope: string;
-    signingAlgorithm: string;
-    storageLabelClaim: string;
-    storageQuotaClaim: string;
-  };
-  passwordLogin: {
-    enabled: boolean;
-  };
-  storageTemplate: {
-    enabled: boolean;
-    hashVerificationEnabled: boolean;
-    template: string;
-  };
-  image: {
-    thumbnailFormat: ImageFormat;
-    thumbnailSize: number;
-    previewFormat: ImageFormat;
-    previewSize: number;
-    quality: number;
-    colorspace: Colorspace;
-    extractEmbedded: boolean;
-  };
-  newVersionCheck: {
-    enabled: boolean;
-  };
-  trash: {
-    enabled: boolean;
-    days: number;
-  };
-  theme: {
-    customCss: string;
-  };
-  library: {
-    scan: {
-      enabled: boolean;
-      cronExpression: string;
-    };
-    watch: {
-      enabled: boolean;
-    };
-  };
-  notifications: {
-    smtp: {
-      enabled: boolean;
-      from: string;
-      replyTo: string;
-      transport: {
-        ignoreCert: boolean;
-        host: string;
-        port: number;
-        username: string;
-        password: string;
-      };
-    };
-  };
-  server: {
-    externalDomain: string;
-    loginPageMessage: string;
-  };
-  user: {
-    deleteDelay: number;
-  };
 }
