+++ server/src/main.ts
@@ -7,8 +7,8 @@ import { existsSync } from 'node:fs';
 import { Worker } from 'node:worker_threads';
 import sirv from 'sirv';
 import { ApiModule, ImmichAdminModule } from 'src/app.module';
+import { LogLevel } from 'src/config';
 import { WEB_ROOT, envName, excludePaths, isDev, serverVersion } from 'src/constants';
-import { LogLevel } from 'src/entities/system-config.entity';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { WebSocketAdapter } from 'src/middleware/websocket.adapter';
 import { ApiService } from 'src/services/api.service';
