+++ server/src/services/media.service.ts
@@ -1,10 +1,5 @@
 import { Inject, Injectable, UnsupportedMediaTypeException } from '@nestjs/common';
 import { dirname } from 'node:path';
-import { GeneratedImageType, StorageCore, StorageFolder } from 'src/cores/storage.core';
-import { SystemConfigCore } from 'src/cores/system-config.core';
-import { SystemConfigFFmpegDto } from 'src/dtos/system-config.dto';
-import { AssetEntity, AssetType } from 'src/entities/asset.entity';
-import { AssetPathType } from 'src/entities/move.entity';
 import {
   AudioCodec,
   Colorspace,
@@ -13,7 +8,12 @@ import {
   TranscodePolicy,
   TranscodeTarget,
   VideoCodec,
-} from 'src/entities/system-config.entity';
+} from 'src/config';
+import { GeneratedImageType, StorageCore, StorageFolder } from 'src/cores/storage.core';
+import { SystemConfigCore } from 'src/cores/system-config.core';
+import { SystemConfigFFmpegDto } from 'src/dtos/system-config.dto';
+import { AssetEntity, AssetType } from 'src/entities/asset.entity';
+import { AssetPathType } from 'src/entities/move.entity';
 import { IAssetRepository, WithoutProperty } from 'src/interfaces/asset.interface';
 import { ICryptoRepository } from 'src/interfaces/crypto.interface';
 import {
